+++ lib/profile_routes.dart
@@ -1,6 +1,5 @@
-import 'package:sint/sint.dart';
-
 import 'package:neom_core/utils/constants/app_route_constants.dart';
+import 'package:sint/sint.dart';
 
 import 'ui/follows/followers_list_page.dart';
 import 'ui/follows/following_list_page.dart';