+++ lib/ui/tabs/profile_items.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:sint/sint.dart';
 import 'package:neom_commons/ui/widgets/images/handled_cached_network_image.dart';
 import 'package:neom_commons/ui/widgets/rating_heart_bar.dart';
 import 'package:neom_commons/utils/app_utilities.dart';
@@ -9,6 +8,7 @@ import 'package:neom_commons/utils/constants/app_page_id_constants.dart';
 import 'package:neom_commons/utils/constants/translations/common_translation_constants.dart';
 import 'package:neom_commons/utils/mappers/base_item_mapper.dart';
 import 'package:neom_core/domain/model/base_item.dart';
+import 'package:sint/sint.dart';
 
 import '../profile_controller.dart';